Fix landing at correct height when navigating to anchor
Because the anchor gets put at the top of the page when you click on a menu item, and the navigation bar is now always fixed at the top of the screen, the anchor will be behind the navigation bar. To fix this, we move the anchor up higher than the content, so that the content displays below the navigation bar when you click a menu item.